requirements.txt
flask
python-dotenv
google-generativeai
python-docx
werkzeug